requirements.txt
streamlit==1.35.0
tensorflow==2.16.1
keras==3.3.3
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.0
opencv-python==4.10.0.84
Pillow==10.3.0
matplotlib==3.9.0
plotly==5.22.0
seaborn==0.13.2
fpdf2==2.7.9
bcrypt==4.1.3
SQLAlchemy==2.0.30
scipy==1.13.1